fix arrow annotate dialog
